refactor code to enable thread sanitization and handle 1 philo

+++ philo/print.c
@@ -22,14 +22,10 @@ void safe_write(t_philo *philo, char *msg)
 	long long current_time;
 
 	pthread_mutex_lock(&philo->params->write_mutex);
-    current_time = get_elapsed_time((long long)philo->params->start_time);
+    current_time = get_elapsed_time(philo->params->start_time);
 	ft_putnbr(current_time);
 	ft_putstr(" ");
 	ft_putnbr(philo->id);
 	ft_putstr(msg);
-    // ft_putstr(" ");
-	// ft_putnbr(philo->num_meals);
-    // ft_putstr(" meal");
-
 	pthread_mutex_unlock(&philo->params->write_mutex);
 }
